requirements.txt
Flask==2.2.2
protobuf==3.20.*
Werkzeug==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gunicorn==20.1.0
tensorflow==2.3.1
Pillow==9.3.0
flask_cors==3.0.10
Flask-Login==0.6.3